Technomancy and JSON
*JSON-ified all the alchemy recipes
*Started porting technomancy. A lot of features are going to get cut do to technical challenges, but I've got some new ideas I want to implement that should make technomancy a little less disjointed and a lot less buggy

+++ src/main/java/com/Da_Technomancer/crossroads/entity/EntityFlameCore.java
@@ -152,7 +152,7 @@ public class EntityFlameCore extends Entity{
 		}
 
 		if(ticksExisted % 8 == 0){
-			int radius = (int) Math.round(FLAME_VEL * (float) ticksExisted);
+			int radius = Math.round(FLAME_VEL * (float) ticksExisted);
 			BlockPos pos = new BlockPos(posX, posY, posZ);
 			boolean lastAction = maxRadius <= radius;
